Rename: Auditable 클래스의 도메인 변경으로 인한 수정
