+++ src/main/java/com/flab/dduikka/domain/vote/domain/Vote.java
@@ -4,7 +4,7 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
-import com.flab.dduikka.domain.Auditable;
+import com.flab.dduikka.common.domain.Auditable;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
